refactor(chat/server): hoist ChatOptions to shared @types
Move shared adapter config out of the deprecated next-js folder so the
express-js adapter doesn't depend on a module slated for removal.
