+++ src/chat/server/@types.ts
@@ -64,6 +64,48 @@ export interface WebSearchConfig {
 }
 
 // ============================================================================
+// Chat Options (namespaced under `chat` in framework adapters)
+// ============================================================================
+
+export interface ChatOptions {
+	/**
+	 * System prompt for the assistant.
+	 * Can be overridden per-request via `beforeRequest`.
+	 */
+	systemPrompt?: string;
+
+	/**
+	 * Maximum number of tool call steps. Defaults to 5.
+	 */
+	maxSteps?: number;
+
+	/**
+	 * Hook called before each request is forwarded to the WaniWani API.
+	 * - Return void to use defaults.
+	 * - Return an object to override messages, systemPrompt, or sessionId.
+	 * - Throw to reject the request (the error message is returned as JSON).
+	 */
+	beforeRequest?: (
+		context: BeforeRequestContext,
+	) =>
+		| Promise<BeforeRequestResult | undefined>
+		| BeforeRequestResult
+		| undefined;
+
+	/**
+	 * Override the MCP server URL directly, bypassing config resolution.
+	 * Useful for development/testing when pointing to a local MCP server.
+	 */
+	mcpServerUrl?: string;
+
+	/**
+	 * Enable web search as an additional tool alongside MCP tools.
+	 * Pass `true` to enable with defaults, or a config object to restrict domains.
+	 */
+	webSearch?: boolean | WebSearchConfig;
+}
+
+// ============================================================================
 // API Handler Options
 // ============================================================================
